vqa.py
from transformers import ViltProcessor, ViltForQuestionAnswering
from PIL import Image

def ask_question(image_path, question):
    # prepare image + question
    image = Image.open(image_path)

    processor = ViltProcessor.from_pretrained("dandelin/vilt-b32-finetuned-vqa")
    model = ViltForQuestionAnswering.from_pretrained("dandelin/vilt-b32-finetuned-vqa")

    # prepare inputs
    encoding = processor(image, question, return_tensors="pt")

    # forward pass
    outputs = model(**encoding)
    logits = outputs.logits
    idx = logits.argmax(-1).item()
    answer = model.config.id2label[idx]

    return answer
