requirements.txt
accelerate==0.20.3
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.2
Flask==2.3.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psutil==5.9.5
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.1
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
